Couple of minor refactors
