+++ web/src/lib/services/map/map.svelte.ts
@@ -1,39 +1,11 @@
 import type {MapPlaceable} from '$lib/interfaces/object';
-import {activeObjectInfo, map} from '$lib/stores/map';
-
-export function getStreetView(lat: number, lng: number) {
-    const streetView = new google.maps.StreetViewService();
-    return streetView
-        .getPanorama({
-            location: new google.maps.LatLng(lat, lng),
-            radius: 30,
-        })
-        .then(({data}: google.maps.StreetViewResponse) => {
-            map.subscribe(value => {
-                const location = data.location!;
-                if (value) {
-                    const pano = value.getStreetView();
-                    pano.setPano(location.pano as string);
-                    pano.setVisible(true);
-                    activeObjectInfo.update(value => ({...value, isMinimized: true}));
-                }
-            });
-        });
-}
+import {map} from '$lib/stores/map';
 
 export function setCenter(lat: number, lng: number) {
     map.subscribe(value => {
         if (value) {
             value.setCenter(new google.maps.LatLng(lat, lng));
             value.setZoom(16);
-        }
-    });
-}
-
-export function setDraggable(isDraggable: boolean) {
-    map.subscribe(value => {
-        if (value) {
-            value.set('draggable', isDraggable);
         }
     });
 }
