+++ web/src/lib/services/map/marker.svelte.ts
@@ -1,0 +1,9 @@
+import {map} from '$lib/stores/map';
+
+export function setDraggable(isDraggable: boolean) {
+    map.subscribe(value => {
+        if (value) {
+            value.set('draggable', isDraggable);
+        }
+    });
+}
